Working synthetic class, split out files
